WAS Robot Rev.1

+++ was_sensor/src/lineTracker_sensor.cpp
@@ -1,5 +1,5 @@
 #include "ros/ros.h"
-#include <std_msgs/UInt32.h>
+#include <std_msgs/UInt8.h>
 #include "BlackLib/BlackLib.h"
 #include <sstream>
 #include <bitset>
@@ -15,7 +15,7 @@ int main(int argc, char **argv)
 {
         ros::init(argc, argv, "was_lineTracker_sensor");
         ros::NodeHandle n;
-        ros::Publisher lineTracker_pub = n.advertise<std_msgs::UInt32>("was_sensor/lineTracking", 1000);
+        ros::Publisher lineTracker_pub = n.advertise<std_msgs::UInt8>("was_sensor/lineTracking", 1000);
         ros::Rate loop_rate(10);
 
         std::bitset<5> trackingValue_bitset;
@@ -25,7 +25,7 @@ int main(int argc, char **argv)
         BlackLib::BlackGPIO t4Value(BlackLib::GPIO_44, BlackLib::input);
         BlackLib::BlackGPIO t5Value(BlackLib::GPIO_26, BlackLib::input);
 
-        std_msgs::UInt32 trackingValue;
+        std_msgs::UInt8 trackingValue;
 
         while (ros::ok()) {
                 trackingValue_bitset[T1] = t1Value.getNumericValue();
